+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - [feat:support setting load balancing strategy per service.](https://github.com/Tencent/spring-cloud-tencent/pull/1701)
 - [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1702)
 - [feat:support traffic mirroring.](https://github.com/Tencent/spring-cloud-tencent/pull/1703)
+- [feat: support custom quickstart circuitbreak delay time.](https://github.com/Tencent/spring-cloud-tencent/pull/1704)